requirements.txt
# Data manipulation and analysis
numpy           # pip install numpy
pandas          # pip install pandas
pandas-datareader # pip install pandas-datareader

# Data visualization
matplotlib      # pip install matplotlib
seaborn         # pip install seaborn
plotly          # pip install plotly

# Machine learning and time series analysis
scikit-learn    # pip install scikit-learn
tensorflow      # pip install tensorflow
statsmodels     # pip install statsmodels

# Backtesting and finance-specific libraries
TA-Lib          # pip install TA-Lib
backtrader      # pip install backtrader

# API requests
requests        # pip install requests

# Environment variables
python-dotenv   # pip install python-dotenv

# Jupyter Notebook
jupyter         # pip install jupyter

# General utilities
tqdm            # pip install tqdm


# To install all the required packages listed in the requirements.txt file, run the following command: pip install -r requirements.txt
